requirements.txt
python-telegram-bot>=20.7
anthropic>=0.40.0
icalendar>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pyyaml>=6.0
pytz>=2024.1
tavily-python>=0.5.0
